--- a/src/centrality/pagerank.jl
+++ b/src/centrality/pagerank.jl
@@ -5,14 +5,12 @@
     pagerank(g, α=0.85, n=100, ϵ=1.0e-6)
 
 Calculate the [PageRank](https://en.wikipedia.org/wiki/PageRank) of the
-directed graph `g` parameterized by damping factor `α`, number of
-iterations `n`, and convergence threshold `ϵ`. Return a vector representing
-the centrality calculated for each node in `g`, or an error if convergence
+graph `g` parameterized by damping factor `α`, number of iterations 
+`n`, and convergence threshold `ϵ`. Return a vector representing the
+centrality calculated for each node in `g`, or an error if convergence
 is not reached within `n` iterations.
 """
-function pagerank end
-
-@traitfn function pagerank(g::AbstractGraph, α=0.85, n=100::Integer, ϵ=1.0e-6)
+function pagerank(g::AbstractGraph, α=0.85, n=100::Integer, ϵ=1.0e-6)
     # collect dangling nodes
     dangling_nodes = [v for v in vertices(g) if outdegree(g, v) == 0]
     N = Int(nv(g))