@sbromberger 2017-10-23
Hi @afternone - thanks for this. I've been out of town for a few days but am getting back. I'll take a look in the next couple days, but paging @jpfairbanks as well.