@sbromberger 2017-11-14
Can we get this ready for merge?