@sbromberger 2017-12-03
I'm going to merge this given #792's promise to get code coverage back up. @jpfairbanks - if we're not at 100% after both PRs are merged, would you please help w/ codecov?